Moved code from example to library.
