+++ example/lib/main.dart
@@ -1,5 +1,5 @@
-import 'package:example/window_paint.dart';
-import 'package:example/window_paint_controller.dart';
+import 'package:window_paint/window_paint.dart';
+import 'package:window_paint/window_paint_controller.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
