+++ example/lib/window_paint_control.dart
@@ -1,6 +1,6 @@
-import 'package:example/custom_radio.dart';
-import 'package:example/custom_radio_controller.dart';
-import 'package:example/window_paint_controller.dart';
+import 'package:window_paint/custom_radio.dart';
+import 'package:window_paint/custom_radio_controller.dart';
+import 'package:window_paint/window_paint_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
