+++ lib/draw/adapters/draw_pencil_adapter.dart
@@ -1,8 +1,8 @@
 import 'dart:ui';
 
-import 'package:example/draw/draw_object_adapter.dart';
-import 'package:example/draw/draw_point.dart';
-import 'package:example/draw/objects/draw_pencil.dart';
+import 'package:window_paint/draw/draw_object_adapter.dart';
+import 'package:window_paint/draw/draw_point.dart';
+import 'package:window_paint/draw/objects/draw_pencil.dart';
 
 class DrawPencilAdapter extends DrawObjectAdapter<DrawPencil> {
   const DrawPencilAdapter();
