+++ lib/draw/objects/draw_rectangle_cross.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
-import 'package:example/draw/draw_point.dart';
-import 'package:example/draw/objects/draw_rectangle.dart';
+import 'package:window_paint/draw/draw_point.dart';
+import 'package:window_paint/draw/objects/draw_rectangle.dart';
 import 'package:flutter/foundation.dart';
 
 class DrawRectangleCross extends DrawRectangle {
