+++ lib/window_paint_mode.dart
