--- test/window_paint_test.dart
@@ -1,13 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:window_paint/window_paint.dart';
-
-void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
-    expect(() => calculator.addOne(null), throwsNoSuchMethodError);
-  });
-}
